Fill in the blanks.
most of the algorithm was learned from esutils.py

The module name is pretty obvious we only to TAN, so
no need to assert.

Avoided depending on fits header.
The current keyword arguments are probably not going to cause big
troubles.

Avoided checking image boundarys. the WCS specs didn't give a limit
on how far the transformation goes. the edge pruning shall be done
by the caller.