add the infrastructure to migrate catalogue schema.
update select_elgs.py to use DECAM_EXTINCTION (EDR3), rather than EXTINCITION

EDR extinction field will be automatically renamed to DECAM_EXTINCTION.

Maybe this is a complexity that we do not want to deal with. but being able
to run the same analysis code on both data sets is kind of cool.