Do not apply extra confidence cut in during export.
The confidence is applied only if a magnitude is required.

The nersc scripts still needs to be updated, as
export-text has gained an additional first argument for the
type of target. (see runtests.sh)

The customize target definitions shall be simplified, to
remove those extra .bands and .color_bands fields. These
are now automatcially inferred from the semantic target definitions.